Add shared logging context utilities
