+++ analysis/analyse.py
@@ -610,7 +610,8 @@ def main(argv: Union[List[str], None] = None) -> None:
 
         if args.output.lower().endswith(".html"):
             logger.info("Generating HTML analysis report...")
-            analyse_logs(args.logs, args.output)
+            log_arg = args.logs[0] if len(args.logs) == 1 else args.logs
+            analyse_logs(log_arg, args.output)
             logger.info(f"✅ Flight analysis completed successfully")
         else:
             if len(args.logs) != 1:
